vm: move vm CLI to cli/vm package
