+++ cli/vm/cli.go
@@ -1,4 +1,4 @@
-package cli
+package vm
 
 import (
 	"bytes"
