+++ cli/vm/vm.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/chzyer/readline"
 	"github.com/nspcc-dev/neo-go/cli/cmdargs"
-	vmcli "github.com/nspcc-dev/neo-go/pkg/vm/cli"
 	"github.com/urfave/cli"
 )
 
@@ -23,6 +22,6 @@ func startVMPrompt(ctx *cli.Context) error {
 	if err := cmdargs.EnsureNone(ctx); err != nil {
 		return err
 	}
-	p := vmcli.NewWithConfig(true, os.Exit, &readline.Config{})
+	p := NewWithConfig(true, os.Exit, &readline.Config{})
 	return p.Run()
 }
